you can now change the float type with features

+++ src/collision/plane.rs
@@ -1,7 +1,8 @@
 use crate::vec::Vector;
+use crate::Float;
 
 pub struct Plane
 {
     normal : Vector,
-    distance_from_origin : f64
+    distance_from_origin : Float
 }
