+++ tests/quadtree.rs
@@ -1,4 +1,5 @@
 use vecto_rs::{spatial::QuadTree, linear::*};
+use vecto_rs::Float;
 
 #[test]
 fn quadtree_simple()
@@ -7,7 +8,7 @@ fn quadtree_simple()
     
     for i in 0..100
     {
-        tree.add((i % 2) == 0, Vector::new2((i % 500) as f64, (i % 500) as f64));
+        tree.add((i % 2) == 0, Vector::new2((i % 500) as Float, (i % 500) as Float));
     }
 
     let cell = tree.query(Vector::new2(230.0, 230.0));
@@ -20,7 +21,7 @@ fn quadtree_simple()
 
     for i in 0..100
     {
-        tree.remove(Vector::new2((i % 500) as f64, (i % 500) as f64));
+        tree.remove(Vector::new2((i % 500) as Float, (i % 500) as Float));
     }
 
     tree.prune();
